pyproject.toml
[tool.poetry]
name = "pike"
version = "0.1.0"
description = ""
authors = ["skelmis <skelmis.craft@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
jinja2 = "^3.1.4"
markdown-it-py = "^3.0.0"
python-frontmatter = "^1.1.0"
typer = "^0.12.4"
skelmis-commons = "^1.3.0"
skelmis-docx = "^1.4.0"


[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
pytest = "^8.3.3"


[tool.poetry.group.docs.dependencies]
sphinx = "^8.0.2"
furo = "^2024.8.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
